chore: install react navigation and imports sort plugin (#171)
* exclude environment, coverage & package-lock files

* install react navigation and imports sort plugin
