+++ .prettierrc.js
@@ -1,7 +1,12 @@
 module.exports = {
   arrowParens: 'avoid',
   bracketSameLine: true,
-  bracketSpacing: false,
+  bracketSpacing: true,
   singleQuote: true,
   trailingComma: 'all',
+  printWidth: 100,
+  jsxBracketSameLine: true,
+  importOrder: ['^@(.*)/(.*)$', '^screens/(.*)$', '^constants/(.*)$', '^[./*]'],
+  importOrderSeparation: true,
+  importOrderSortSpecifiers: true,
 };
