+++ android/app/src/main/java/com/reactnativebase/MainActivity.java
@@ -3,6 +3,7 @@ package com.reactnativebase;
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
 import com.facebook.react.ReactRootView;
+import android.os.Bundle; // React Navigation requirement
 
 public class MainActivity extends ReactActivity {
 
@@ -22,6 +23,14 @@ public class MainActivity extends ReactActivity {
   @Override
   protected ReactActivityDelegate createReactActivityDelegate() {
     return new MainActivityDelegate(this, getMainComponentName());
+  }
+
+  /**
+   * React Navigation specific code
+   */
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    super.onCreate(null);
   }
 
   public static class MainActivityDelegate extends ReactActivityDelegate {
