+++ babel.config.js
@@ -1,3 +1,25 @@
 module.exports = {
   presets: ['module:metro-react-native-babel-preset'],
+  plugins: [
+    [
+      'module-resolver',
+      {
+        root: ['./src'],
+        extensions: [
+          '.ios.ts',
+          '.android.ts',
+          '.ts',
+          '.ios.tsx',
+          '.android.tsx',
+          '.jsx',
+          '.js',
+          '.json',
+        ],
+        alias: {
+          screens: './src/screens',
+          constants: './src/constants',
+        },
+      },
+    ],
+  ],
 };
