+++ index.js
@@ -1,9 +1,6 @@
-/**
- * @format
- */
+import { AppRegistry } from 'react-native';
 
-import {AppRegistry} from 'react-native';
 import App from './App';
-import {name as appName} from './app.json';
+import { name as appName } from './app.json';
 
 AppRegistry.registerComponent(appName, () => App);
