+++ src/constants/screens.ts
@@ -1,0 +1,1 @@
+export const WELCOME_SCREEN = 'WelcomeScreen';
